requirements.txt
Flask==3.0.0
moviepy==1.0.3
groq==0.4.1
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
